requirements.txt
google-cloud-vertexai>=1.35.0
google-generativeai>=0.3.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
